Refactor test namespaces to include Sut folder
Updated test class namespaces from Moda.Planning.Application.Tests to Moda.Planning.Application.Tests.Sut for better organization. Also reordered and merged some using directives. No changes to test logic.
