+++ Moda.Services/Moda.Planning/tests/Moda.Planning.Application.Tests/Sut/EstimationScales/Commands/DeleteEstimationScaleCommandHandlerTests.cs
@@ -1,27 +1,27 @@
-using Microsoft.Extensions.Logging;
+﻿using Microsoft.Extensions.Logging;
 using Moda.Planning.Application.EstimationScales.Commands;
 using Moda.Planning.Application.Tests.Infrastructure;
 using Moda.Planning.Domain.Tests.Data;
 using Moq;
 
-namespace Moda.Planning.Application.Tests.EstimationScales.Commands;
+namespace Moda.Planning.Application.Tests.Sut.EstimationScales.Commands;
 
-public class SetEstimationScaleActiveStatusCommandHandlerTests : IDisposable
+public class DeleteEstimationScaleCommandHandlerTests : IDisposable
 {
     private readonly FakePlanningDbContext _dbContext;
-    private readonly SetEstimationScaleActiveStatusCommandHandler _handler;
-    private readonly Mock<ILogger<SetEstimationScaleActiveStatusCommandHandler>> _mockLogger;
+    private readonly DeleteEstimationScaleCommandHandler _handler;
+    private readonly Mock<ILogger<DeleteEstimationScaleCommandHandler>> _mockLogger;
 
-    public SetEstimationScaleActiveStatusCommandHandlerTests()
+    public DeleteEstimationScaleCommandHandlerTests()
     {
         _dbContext = new FakePlanningDbContext();
-        _mockLogger = new Mock<ILogger<SetEstimationScaleActiveStatusCommandHandler>>();
+        _mockLogger = new Mock<ILogger<DeleteEstimationScaleCommandHandler>>();
 
-        _handler = new SetEstimationScaleActiveStatusCommandHandler(_dbContext, _mockLogger.Object);
+        _handler = new DeleteEstimationScaleCommandHandler(_dbContext, _mockLogger.Object);
     }
 
     [Fact]
-    public async Task Handle_ShouldDeactivate_WhenOtherActiveScalesExist()
+    public async Task Handle_ShouldDeleteScale_WhenScaleExistsAndNotInUse()
     {
         // Arrange
         var scale = new EstimationScaleFaker().WithId(1).Generate();
@@ -29,32 +29,32 @@ public class SetEstimationScaleActiveStatusCommandHandlerTests : IDisposable
         _dbContext.AddEstimationScale(scale);
         _dbContext.AddEstimationScale(otherScale);
 
-        var command = new SetEstimationScaleActiveStatusCommand(scale.Id, false);
+        var command = new DeleteEstimationScaleCommand(scale.Id);
 
         // Act
         var result = await _handler.Handle(command, TestContext.Current.CancellationToken);
 
         // Assert
         result.IsSuccess.Should().BeTrue();
-        scale.IsActive.Should().BeFalse();
         _dbContext.SaveChangesCallCount.Should().Be(1);
     }
 
     [Fact]
-    public async Task Handle_ShouldActivate_WhenScaleIsInactive()
+    public async Task Handle_ShouldDeleteInactiveScale_WhenActiveScalesExist()
     {
         // Arrange
         var inactiveScale = new EstimationScaleFaker().WithId(1).WithIsActive(false).Generate();
+        var activeScale = new EstimationScaleFaker().WithId(2).Generate();
         _dbContext.AddEstimationScale(inactiveScale);
+        _dbContext.AddEstimationScale(activeScale);
 
-        var command = new SetEstimationScaleActiveStatusCommand(inactiveScale.Id, true);
+        var command = new DeleteEstimationScaleCommand(inactiveScale.Id);
 
         // Act
         var result = await _handler.Handle(command, TestContext.Current.CancellationToken);
 
         // Assert
         result.IsSuccess.Should().BeTrue();
-        inactiveScale.IsActive.Should().BeTrue();
         _dbContext.SaveChangesCallCount.Should().Be(1);
     }
 
@@ -62,7 +62,7 @@ public class SetEstimationScaleActiveStatusCommandHandlerTests : IDisposable
     public async Task Handle_ShouldFail_WhenScaleDoesNotExist()
     {
         // Arrange
-        var command = new SetEstimationScaleActiveStatusCommand(999, false);
+        var command = new DeleteEstimationScaleCommand(999);
 
         // Act
         var result = await _handler.Handle(command, TestContext.Current.CancellationToken);
@@ -74,7 +74,30 @@ public class SetEstimationScaleActiveStatusCommandHandlerTests : IDisposable
     }
 
     [Fact]
-    public async Task Handle_ShouldFail_WhenDeactivatingLastActiveScale()
+    public async Task Handle_ShouldFail_WhenScaleIsInUseByPokerSession()
+    {
+        // Arrange
+        var scale = new EstimationScaleFaker().WithId(1).Generate();
+        var otherScale = new EstimationScaleFaker().WithId(2).Generate();
+        _dbContext.AddEstimationScale(scale);
+        _dbContext.AddEstimationScale(otherScale);
+
+        var session = new PokerSessionFaker().WithEstimationScaleId(scale.Id).Generate();
+        _dbContext.AddPokerSession(session);
+
+        var command = new DeleteEstimationScaleCommand(scale.Id);
+
+        // Act
+        var result = await _handler.Handle(command, TestContext.Current.CancellationToken);
+
+        // Assert
+        result.IsFailure.Should().BeTrue();
+        result.Error.Should().Contain("in use");
+        _dbContext.SaveChangesCallCount.Should().Be(0);
+    }
+
+    [Fact]
+    public async Task Handle_ShouldFail_WhenDeletingLastActiveScale()
     {
         // Arrange
         var scale = new EstimationScaleFaker().WithId(1).Generate();
@@ -82,7 +105,7 @@ public class SetEstimationScaleActiveStatusCommandHandlerTests : IDisposable
         _dbContext.AddEstimationScale(scale);
         _dbContext.AddEstimationScale(inactiveScale);
 
-        var command = new SetEstimationScaleActiveStatusCommand(scale.Id, false);
+        var command = new DeleteEstimationScaleCommand(scale.Id);
 
         // Act
         var result = await _handler.Handle(command, TestContext.Current.CancellationToken);
@@ -90,45 +113,6 @@ public class SetEstimationScaleActiveStatusCommandHandlerTests : IDisposable
         // Assert
         result.IsFailure.Should().BeTrue();
         result.Error.Should().Contain("last active");
-        scale.IsActive.Should().BeTrue();
-        _dbContext.SaveChangesCallCount.Should().Be(0);
-    }
-
-    [Fact]
-    public async Task Handle_ShouldFail_WhenAlreadyActive()
-    {
-        // Arrange
-        var scale = new EstimationScaleFaker().WithId(1).Generate();
-        _dbContext.AddEstimationScale(scale);
-
-        var command = new SetEstimationScaleActiveStatusCommand(scale.Id, true);
-
-        // Act
-        var result = await _handler.Handle(command, TestContext.Current.CancellationToken);
-
-        // Assert
-        result.IsFailure.Should().BeTrue();
-        result.Error.Should().Contain("already active");
-        _dbContext.SaveChangesCallCount.Should().Be(0);
-    }
-
-    [Fact]
-    public async Task Handle_ShouldFail_WhenAlreadyInactive()
-    {
-        // Arrange
-        var inactiveScale = new EstimationScaleFaker().WithId(1).WithIsActive(false).Generate();
-        var otherScale = new EstimationScaleFaker().WithId(2).Generate();
-        _dbContext.AddEstimationScale(inactiveScale);
-        _dbContext.AddEstimationScale(otherScale);
-
-        var command = new SetEstimationScaleActiveStatusCommand(inactiveScale.Id, false);
-
-        // Act
-        var result = await _handler.Handle(command, TestContext.Current.CancellationToken);
-
-        // Assert
-        result.IsFailure.Should().BeTrue();
-        result.Error.Should().Contain("already inactive");
         _dbContext.SaveChangesCallCount.Should().Be(0);
     }
