+++ Moda.Services/Moda.Planning/tests/Moda.Planning.Application.Tests/Sut/EstimationScales/Commands/UpdateEstimationScaleCommandHandlerTests.cs
@@ -1,11 +1,11 @@
-using Microsoft.Extensions.Logging;
+﻿using Microsoft.Extensions.Logging;
 using Moda.Planning.Application.EstimationScales.Commands;
 using Moda.Planning.Application.Tests.Infrastructure;
 using Moda.Planning.Domain.Models.PlanningPoker;
 using Moda.Planning.Domain.Tests.Data;
 using Moq;
 
-namespace Moda.Planning.Application.Tests.EstimationScales.Commands;
+namespace Moda.Planning.Application.Tests.Sut.EstimationScales.Commands;
 
 public class UpdateEstimationScaleCommandHandlerTests : IDisposable
 {
