+++ Moda.Services/Moda.Planning/tests/Moda.Planning.Application.Tests/Sut/PokerSessions/Commands/CompletePokerSessionCommandHandlerTests.cs
@@ -1,59 +1,62 @@
 ﻿using Microsoft.Extensions.Logging;
+using Moda.Common.Application.Interfaces;
 using Moda.Planning.Application.PokerSessions.Commands;
 using Moda.Planning.Application.PokerSessions.Interfaces;
 using Moda.Planning.Application.Tests.Infrastructure;
 using Moda.Planning.Domain.Enums;
 using Moda.Planning.Domain.Tests.Data;
 using Moq;
+using NodaTime;
 
-namespace Moda.Planning.Application.Tests.PokerSessions.Commands;
+namespace Moda.Planning.Application.Tests.Sut.PokerSessions.Commands;
 
-public class RemovePokerRoundCommandHandlerTests : IDisposable
+public class CompletePokerSessionCommandHandlerTests : IDisposable
 {
     private readonly FakePlanningDbContext _dbContext;
-    private readonly RemovePokerRoundCommandHandler _handler;
-    private readonly Mock<ILogger<RemovePokerRoundCommandHandler>> _mockLogger;
+    private readonly CompletePokerSessionCommandHandler _handler;
+    private readonly Mock<ILogger<CompletePokerSessionCommandHandler>> _mockLogger;
     private readonly Mock<IPokerSessionNotifier> _mockNotifier;
+    private readonly Mock<IDateTimeProvider> _mockDateTimeProvider;
 
     private readonly PokerSessionFaker _sessionFaker;
 
-    public RemovePokerRoundCommandHandlerTests()
+    public CompletePokerSessionCommandHandlerTests()
     {
         _dbContext = new FakePlanningDbContext();
-        _mockLogger = new Mock<ILogger<RemovePokerRoundCommandHandler>>();
+        _mockLogger = new Mock<ILogger<CompletePokerSessionCommandHandler>>();
         _mockNotifier = new Mock<IPokerSessionNotifier>();
+        _mockDateTimeProvider = new Mock<IDateTimeProvider>();
+        _mockDateTimeProvider.Setup(d => d.Now).Returns(Instant.FromUtc(2026, 1, 15, 10, 0));
 
-        _handler = new RemovePokerRoundCommandHandler(_dbContext, _mockNotifier.Object, _mockLogger.Object);
+        _handler = new CompletePokerSessionCommandHandler(_dbContext, _mockDateTimeProvider.Object, _mockNotifier.Object, _mockLogger.Object);
         _sessionFaker = new PokerSessionFaker();
     }
 
     [Fact]
-    public async Task Handle_ShouldRemoveRound_WhenRoundExists()
+    public async Task Handle_ShouldCompleteSession_WhenSessionIsActive()
     {
         // Arrange
         var session = _sessionFaker.WithStatus(PokerSessionStatus.Active).Generate();
         _dbContext.AddPokerSession(session);
 
-        session.AddRound("Round to remove");
-        var roundId = session.Rounds.First().Id;
-
-        var command = new RemovePokerRoundCommand(session.Id, roundId);
+        var command = new CompletePokerSessionCommand(session.Id);
 
         // Act
         var result = await _handler.Handle(command, TestContext.Current.CancellationToken);
 
         // Assert
         result.IsSuccess.Should().BeTrue();
-        session.Rounds.Should().BeEmpty();
+        session.Status.Should().Be(PokerSessionStatus.Completed);
+        session.CompletedOn.Should().NotBeNull();
         _dbContext.SaveChangesCallCount.Should().Be(1);
-        _mockNotifier.Verify(n => n.NotifyRoundRemoved(session.Id, roundId), Times.Once);
+        _mockNotifier.Verify(n => n.NotifySessionCompleted(session.Id), Times.Once);
     }
 
     [Fact]
     public async Task Handle_ShouldFail_WhenSessionDoesNotExist()
     {
         // Arrange
-        var command = new RemovePokerRoundCommand(Guid.NewGuid(), Guid.NewGuid());
+        var command = new CompletePokerSessionCommand(Guid.NewGuid());
 
         // Act
         var result = await _handler.Handle(command, TestContext.Current.CancellationToken);
@@ -65,31 +68,13 @@ public class RemovePokerRoundCommandHandlerTests : IDisposable
     }
 
     [Fact]
-    public async Task Handle_ShouldFail_WhenRoundDoesNotExist()
-    {
-        // Arrange
-        var session = _sessionFaker.WithStatus(PokerSessionStatus.Active).Generate();
-        _dbContext.AddPokerSession(session);
-
-        var command = new RemovePokerRoundCommand(session.Id, Guid.NewGuid());
-
-        // Act
-        var result = await _handler.Handle(command, TestContext.Current.CancellationToken);
-
-        // Assert
-        result.IsFailure.Should().BeTrue();
-        result.Error.Should().Contain("not found");
-        _dbContext.SaveChangesCallCount.Should().Be(0);
-    }
-
-    [Fact]
-    public async Task Handle_ShouldFail_WhenSessionIsNotActive()
+    public async Task Handle_ShouldFail_WhenSessionIsAlreadyCompleted()
     {
         // Arrange
         var session = _sessionFaker.WithStatus(PokerSessionStatus.Completed).Generate();
         _dbContext.AddPokerSession(session);
 
-        var command = new RemovePokerRoundCommand(session.Id, Guid.NewGuid());
+        var command = new CompletePokerSessionCommand(session.Id);
 
         // Act
         var result = await _handler.Handle(command, TestContext.Current.CancellationToken);
