+++ Moda.Services/Moda.Planning/tests/Moda.Planning.Application.Tests/Sut/PokerSessions/Commands/RemovePokerRoundCommandHandlerTests.cs
@@ -1,62 +1,59 @@
-using Microsoft.Extensions.Logging;
-using Moda.Common.Application.Interfaces;
+﻿using Microsoft.Extensions.Logging;
 using Moda.Planning.Application.PokerSessions.Commands;
 using Moda.Planning.Application.PokerSessions.Interfaces;
 using Moda.Planning.Application.Tests.Infrastructure;
 using Moda.Planning.Domain.Enums;
 using Moda.Planning.Domain.Tests.Data;
 using Moq;
-using NodaTime;
 
-namespace Moda.Planning.Application.Tests.PokerSessions.Commands;
+namespace Moda.Planning.Application.Tests.Sut.PokerSessions.Commands;
 
-public class CompletePokerSessionCommandHandlerTests : IDisposable
+public class RemovePokerRoundCommandHandlerTests : IDisposable
 {
     private readonly FakePlanningDbContext _dbContext;
-    private readonly CompletePokerSessionCommandHandler _handler;
-    private readonly Mock<ILogger<CompletePokerSessionCommandHandler>> _mockLogger;
+    private readonly RemovePokerRoundCommandHandler _handler;
+    private readonly Mock<ILogger<RemovePokerRoundCommandHandler>> _mockLogger;
     private readonly Mock<IPokerSessionNotifier> _mockNotifier;
-    private readonly Mock<IDateTimeProvider> _mockDateTimeProvider;
 
     private readonly PokerSessionFaker _sessionFaker;
 
-    public CompletePokerSessionCommandHandlerTests()
+    public RemovePokerRoundCommandHandlerTests()
     {
         _dbContext = new FakePlanningDbContext();
-        _mockLogger = new Mock<ILogger<CompletePokerSessionCommandHandler>>();
+        _mockLogger = new Mock<ILogger<RemovePokerRoundCommandHandler>>();
         _mockNotifier = new Mock<IPokerSessionNotifier>();
-        _mockDateTimeProvider = new Mock<IDateTimeProvider>();
-        _mockDateTimeProvider.Setup(d => d.Now).Returns(Instant.FromUtc(2026, 1, 15, 10, 0));
 
-        _handler = new CompletePokerSessionCommandHandler(_dbContext, _mockDateTimeProvider.Object, _mockNotifier.Object, _mockLogger.Object);
+        _handler = new RemovePokerRoundCommandHandler(_dbContext, _mockNotifier.Object, _mockLogger.Object);
         _sessionFaker = new PokerSessionFaker();
     }
 
     [Fact]
-    public async Task Handle_ShouldCompleteSession_WhenSessionIsActive()
+    public async Task Handle_ShouldRemoveRound_WhenRoundExists()
     {
         // Arrange
         var session = _sessionFaker.WithStatus(PokerSessionStatus.Active).Generate();
         _dbContext.AddPokerSession(session);
 
-        var command = new CompletePokerSessionCommand(session.Id);
+        session.AddRound("Round to remove");
+        var roundId = session.Rounds.First().Id;
+
+        var command = new RemovePokerRoundCommand(session.Id, roundId);
 
         // Act
         var result = await _handler.Handle(command, TestContext.Current.CancellationToken);
 
         // Assert
         result.IsSuccess.Should().BeTrue();
-        session.Status.Should().Be(PokerSessionStatus.Completed);
-        session.CompletedOn.Should().NotBeNull();
+        session.Rounds.Should().BeEmpty();
         _dbContext.SaveChangesCallCount.Should().Be(1);
-        _mockNotifier.Verify(n => n.NotifySessionCompleted(session.Id), Times.Once);
+        _mockNotifier.Verify(n => n.NotifyRoundRemoved(session.Id, roundId), Times.Once);
     }
 
     [Fact]
     public async Task Handle_ShouldFail_WhenSessionDoesNotExist()
     {
         // Arrange
-        var command = new CompletePokerSessionCommand(Guid.NewGuid());
+        var command = new RemovePokerRoundCommand(Guid.NewGuid(), Guid.NewGuid());
 
         // Act
         var result = await _handler.Handle(command, TestContext.Current.CancellationToken);
@@ -68,13 +65,31 @@ public class CompletePokerSessionCommandHandlerTests : IDisposable
     }
 
     [Fact]
-    public async Task Handle_ShouldFail_WhenSessionIsAlreadyCompleted()
+    public async Task Handle_ShouldFail_WhenRoundDoesNotExist()
+    {
+        // Arrange
+        var session = _sessionFaker.WithStatus(PokerSessionStatus.Active).Generate();
+        _dbContext.AddPokerSession(session);
+
+        var command = new RemovePokerRoundCommand(session.Id, Guid.NewGuid());
+
+        // Act
+        var result = await _handler.Handle(command, TestContext.Current.CancellationToken);
+
+        // Assert
+        result.IsFailure.Should().BeTrue();
+        result.Error.Should().Contain("not found");
+        _dbContext.SaveChangesCallCount.Should().Be(0);
+    }
+
+    [Fact]
+    public async Task Handle_ShouldFail_WhenSessionIsNotActive()
     {
         // Arrange
         var session = _sessionFaker.WithStatus(PokerSessionStatus.Completed).Generate();
         _dbContext.AddPokerSession(session);
 
-        var command = new CompletePokerSessionCommand(session.Id);
+        var command = new RemovePokerRoundCommand(session.Id, Guid.NewGuid());
 
         // Act
         var result = await _handler.Handle(command, TestContext.Current.CancellationToken);
