+++ Moda.Services/Moda.Planning/tests/Moda.Planning.Application.Tests/Sut/PokerSessions/Commands/ResetPokerRoundCommandHandlerTests.cs
@@ -1,5 +1,4 @@
-using Microsoft.Extensions.Logging;
-using Moda.Common.Application.Interfaces;
+﻿using Microsoft.Extensions.Logging;
 using Moda.Planning.Application.PokerSessions.Commands;
 using Moda.Planning.Application.PokerSessions.Interfaces;
 using Moda.Planning.Application.Tests.Infrastructure;
@@ -8,33 +7,29 @@ using Moda.Planning.Domain.Tests.Data;
 using Moq;
 using NodaTime;
 
-namespace Moda.Planning.Application.Tests.PokerSessions.Commands;
+namespace Moda.Planning.Application.Tests.Sut.PokerSessions.Commands;
 
-public class WithdrawVoteCommandHandlerTests : IDisposable
+public class ResetPokerRoundCommandHandlerTests : IDisposable
 {
     private readonly FakePlanningDbContext _dbContext;
-    private readonly WithdrawVoteCommandHandler _handler;
-    private readonly Mock<ILogger<WithdrawVoteCommandHandler>> _mockLogger;
+    private readonly ResetPokerRoundCommandHandler _handler;
+    private readonly Mock<ILogger<ResetPokerRoundCommandHandler>> _mockLogger;
     private readonly Mock<IPokerSessionNotifier> _mockNotifier;
-    private readonly Mock<ICurrentUser> _mockCurrentUser;
 
     private readonly PokerSessionFaker _sessionFaker;
-    private readonly string _currentUserId = Guid.NewGuid().ToString();
 
-    public WithdrawVoteCommandHandlerTests()
+    public ResetPokerRoundCommandHandlerTests()
     {
         _dbContext = new FakePlanningDbContext();
-        _mockLogger = new Mock<ILogger<WithdrawVoteCommandHandler>>();
+        _mockLogger = new Mock<ILogger<ResetPokerRoundCommandHandler>>();
         _mockNotifier = new Mock<IPokerSessionNotifier>();
-        _mockCurrentUser = new Mock<ICurrentUser>();
-        _mockCurrentUser.Setup(u => u.GetUserId()).Returns(_currentUserId);
 
-        _handler = new WithdrawVoteCommandHandler(_dbContext, _mockCurrentUser.Object, _mockNotifier.Object, _mockLogger.Object);
+        _handler = new ResetPokerRoundCommandHandler(_dbContext, _mockNotifier.Object, _mockLogger.Object);
         _sessionFaker = new PokerSessionFaker();
     }
 
     [Fact]
-    public async Task Handle_ShouldWithdrawVote_WhenVoteExists()
+    public async Task Handle_ShouldResetRound_WhenRoundIsRevealed()
     {
         // Arrange
         var session = _sessionFaker.WithStatus(PokerSessionStatus.Active).Generate();
@@ -42,22 +37,24 @@ public class WithdrawVoteCommandHandlerTests : IDisposable
 
         session.AddRound("Story");
         var round = session.Rounds.First();
-        session.SubmitVote(round.Id, _currentUserId, "8", Instant.FromUtc(2026, 1, 15, 9, 0));
+        session.SubmitVote(round.Id, Guid.NewGuid().ToString(), "5", Instant.FromUtc(2026, 1, 15, 10, 0));
+        session.RevealRound(round.Id);
 
-        var command = new WithdrawVoteCommand(session.Id, round.Id);
+        var command = new ResetPokerRoundCommand(session.Id, round.Id);
 
         // Act
         var result = await _handler.Handle(command, TestContext.Current.CancellationToken);
 
         // Assert
         result.IsSuccess.Should().BeTrue();
+        round.Status.Should().Be(PokerRoundStatus.Voting);
         round.Votes.Should().BeEmpty();
         _dbContext.SaveChangesCallCount.Should().Be(1);
-        _mockNotifier.Verify(n => n.NotifyVoteWithdrawn(session.Id, round.Id, _currentUserId), Times.Once);
+        _mockNotifier.Verify(n => n.NotifyRoundReset(session.Id, round.Id), Times.Once);
     }
 
     [Fact]
-    public async Task Handle_ShouldFail_WhenNoVoteExists()
+    public async Task Handle_ShouldResetRound_WhenRoundIsVoting()
     {
         // Arrange
         var session = _sessionFaker.WithStatus(PokerSessionStatus.Active).Generate();
@@ -66,22 +63,23 @@ public class WithdrawVoteCommandHandlerTests : IDisposable
         session.AddRound("Story");
         var round = session.Rounds.First();
 
-        var command = new WithdrawVoteCommand(session.Id, round.Id);
+
+        var command = new ResetPokerRoundCommand(session.Id, round.Id);
 
         // Act
         var result = await _handler.Handle(command, TestContext.Current.CancellationToken);
 
         // Assert
-        result.IsFailure.Should().BeTrue();
-        result.Error.Should().Contain("No vote found");
-        _dbContext.SaveChangesCallCount.Should().Be(0);
+        result.IsSuccess.Should().BeTrue();
+        round.Status.Should().Be(PokerRoundStatus.Voting);
+        _dbContext.SaveChangesCallCount.Should().Be(1);
     }
 
     [Fact]
     public async Task Handle_ShouldFail_WhenSessionDoesNotExist()
     {
         // Arrange
-        var command = new WithdrawVoteCommand(Guid.NewGuid(), Guid.NewGuid());
+        var command = new ResetPokerRoundCommand(Guid.NewGuid(), Guid.NewGuid());
 
         // Act
         var result = await _handler.Handle(command, TestContext.Current.CancellationToken);
@@ -89,6 +87,23 @@ public class WithdrawVoteCommandHandlerTests : IDisposable
         // Assert
         result.IsFailure.Should().BeTrue();
         result.Error.Should().Contain("not found");
+        _dbContext.SaveChangesCallCount.Should().Be(0);
+    }
+
+    [Fact]
+    public async Task Handle_ShouldFail_WhenRoundDoesNotExist()
+    {
+        // Arrange
+        var session = _sessionFaker.WithStatus(PokerSessionStatus.Active).Generate();
+        _dbContext.AddPokerSession(session);
+
+        var command = new ResetPokerRoundCommand(session.Id, Guid.NewGuid());
+
+        // Act
+        var result = await _handler.Handle(command, TestContext.Current.CancellationToken);
+
+        // Assert
+        result.IsFailure.Should().BeTrue();
         _dbContext.SaveChangesCallCount.Should().Be(0);
     }
