+++ Moda.Services/Moda.Planning/tests/Moda.Planning.Application.Tests/Sut/PokerSessions/Commands/SubmitVoteCommandHandlerTests.cs
@@ -1,4 +1,4 @@
-using Microsoft.Extensions.Logging;
+﻿using Microsoft.Extensions.Logging;
 using Moda.Common.Application.Interfaces;
 using Moda.Planning.Application.PokerSessions.Commands;
 using Moda.Planning.Application.PokerSessions.Interfaces;
@@ -8,7 +8,7 @@ using Moda.Planning.Domain.Tests.Data;
 using Moq;
 using NodaTime;
 
-namespace Moda.Planning.Application.Tests.PokerSessions.Commands;
+namespace Moda.Planning.Application.Tests.Sut.PokerSessions.Commands;
 
 public class SubmitVoteCommandHandlerTests : IDisposable
 {
